fix last submit display
